compilation on solaris correction

+++ src/IclModel.cpp
@@ -465,7 +465,7 @@ List IclModel::greedy_merge_path(){
   int nbmerge = 0;
   List obs_stats = this->get_obs_stats();
   arma::vec counts =as<arma::vec>(obs_stats["counts"]);
-  double icl = this->icl_emiss(obs_stats)-log(K)+arma::accu(lgamma(counts))-lgamma(N);
+  double icl = this->icl_emiss(obs_stats)-log(static_cast<double>(K))+arma::accu(lgamma(counts))-lgamma(N);
   double iclold = icl;
   int k=0;
   int l=0;
@@ -490,7 +490,7 @@ List IclModel::greedy_merge_path(){
     obs_stats = this->get_obs_stats();
     if(merge_mat.getValue()>-std::numeric_limits<double>::infinity()){
       counts =as<arma::vec>(obs_stats["counts"]);
-      icl = this->icl_emiss(obs_stats)-log(K)+arma::accu(lgamma(counts))-lgamma(N);
+      icl = this->icl_emiss(obs_stats)-log(static_cast<double>(K))+arma::accu(lgamma(counts))-lgamma(N);
     }else{
       icl = -std::numeric_limits<double>::infinity();
     }
